+++ NEWS.md
@@ -1,5 +1,9 @@
 # testthat (development version)
 
+* `expect_snapshot()` gains new `error` argument which controls whether or not
+  an error is expected. If an unexpected error is thrown, or an expected error
+  is not thrown, `expect_snapshot()` will fail (even on CRAN) (#1200).
+
 * `RstudioReporter` has been renamed to `RStudioReporter`.
 
 * `local_reproducible_output()` now sets the LANGUAGE env var to "en". This